% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_accuracy.R
\name{classAccuracy}
\alias{classAccuracy}
\alias{classAccuracy.hsmclass}
\alias{classAccuracy.default}
\title{classAccuracy}
\usage{
classAccuracy(..., conf.level = 0.95)

\method{classAccuracy}{hsmclass}(object, conf.level = 0.95, ...)

\method{classAccuracy}{default}(refdata, predictions, equal.int = NA,
  def.int = NA, conf.level = 0.95, ...)
}
\arguments{
\item{...}{further arguments passed to or used by methods.}

\item{conf.level}{the confidence level used to compute the confidence intervals of the overall accuracy.}

\item{object}{object of class \code{'hsmclass'} created by function \code{\link{HSMclass}}.}

\item{refdata}{\code{vector} containing the values of the continuous
response variable used in the prediction model.}

\item{predictions}{\code{vector} containing the predictions for the response values
of the prediction model.
\strong{Note:}(\code{refdata} and \code{predictions} have to correspond to each other).}

\item{equal.int}{an equidistant class interval to be evaluated as classifciation scheme. Defaults to \code{NA}.
Can only be set if \code{def.int} is \code{NA}.}

\item{def.int}{a \code{vector} defining an arbitrary set of class breaks to be evaluated as classifciation scheme.
Defaults to \code{NA}. Can only be set if \code{def.int} is \code{NA}.}
}
\value{
\code{classAccuracy} returns an object of class \code{"classaccur"}

An object of class \code{"classaccur"} returns a \code{list} of the following components:

\item{rsquared_val}{the coefficient of determination of the prediction model, calulated based on
                     \code{refdata} and \code{predictions} }.
 \item{predictions}{ \code{vector} containing the predictions given to \code{classAccuracy} }
 \item{classwidth}{ \code{vector} containing the class width}
 \item{def.classbreaks}{ \code{vector} containing arbitrary class break values}
 \item{equal.classbreaks}{ \code{vector} containing equidistant class break values}
 \item{overall.accuracy}{the overall accuracy of the classification scheme}
 \item{conf.oaa}{ \code{vector} containing \code{[1]} the lower confidence limit of the OAA,
                                           \code{[2]} the upper confidence limit of the OAA,
                                       and \code{[3]} the confidence level}
 \item{prodaccuracy}{a \code{\link[base]{matrix}} containing the producer's accuracy for each class}
 \item{usersaccuracy}{a \code{\link[base]{matrix}} containing the user's accuracy for each class}
 \item{no.ref.classes}{ a \code{vector} containing the number of reference data for each class}
 \item{cohenskappa}{Cohen's Kappa Coefficient}
 \item{map.accuracy}{the map accuracy}
 \item{Quantity.Disagreement}{ the quantity disagreement}
 \item{Allocation.Disagreement}{the allocation disagreement}
}
\description{
Calculates common classification accuracies measures known from remote sensing
applications based on a Confusion Matrix that compares Reference classes to their
respective class predictions. The computed accuracy measures are:
\itemize{
        \item \emph{Overall Classification Accuracy (OAA)}
        \item \emph{Producer's Accuracy (PA)}
        \item \emph{User's Accuracy (UA)}
        \item \emph{Cohen's Kappa Coefficient (k)}
        \item \emph{Quantity Disagreement}
        \item \emph{Allocation Disagreement}
         }
}
\examples{


#----
# 1.) Example: classification accuracy for equidistant class width of 100:
#----

acc.equal<- classAccuracy(refdata.gr, predictions.gr, equal.int = 100)
summary(acc.equal)


#----
# 2.) Example: classification accuracy for arbitrary class breaks:
#----

acc.def<- classAccuracy(refdata.gr, predictions.gr,
                       def.int = c(0, 150, 200, 430, 610, 880))
summary(acc.def)


#----
# 3.) Example: classification accuracy for optimal class breaks:
#----

# run HSMclass:
\dontrun{
hsm<- HSMclass(refdata.gr, predictions.gr, nclasses = 6,
              iterations = 1000, coolfactor=0.99, InitTemp = 80,
              weight.norefs = 2, weight.classwidth = 2)

# calculate accuracy:
acc.opti<- classAccuracy(hsm)
summary(acc.opti)
}

}
\references{
Congalton,  R.G.;  Green,  K.  \emph{Assessing  the  Accuracy  of  Remotely  Sensed  Data:  Principles  and
Practices}; Lewis Publications: Boca Raton, FL, USA, 1999; p. 137.

Richards, J.A. \emph{Remote Sensing Digital Image Analysis: An Introduction}, 5th ed.; Springer: Berlin, Germany, 2013.

Hill, A., Breschan, J., & Mandallaz, D. (2014). Accuracy assessment of timber
            volume maps using forest inventory data and LiDAR canopy height models.
            \emph{Forests}, \strong{5(9)}, 2253-2275.
}
