% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qml.R
\name{create_qml}
\alias{create_qml}
\alias{create_qml.classaccur}
\title{create_qml}
\usage{
create_qml(x, ...)

\method{create_qml}{classaccur}(x, out.file, pallete.name = "YlOrRd", ...)
}
\arguments{
\item{x}{object of class \code{'classaccur'} containing an evaluated classification scheme.}

\item{...}{additional arguments, so far ignored.}

\item{out.file}{destination-folder were the qml-file should be stored}

\item{pallete.name}{code of color-palette to be used for visualization
(see \code{\link[RColorBrewer]{display.brewer.all}}).
For an overview type \code{display.brewer.all()} into prompt.
Defaults to "YlOrRd".}
}
\description{
Creates a qml-file of a given classification scheme produced by \code{\link{classAccuracy}}
for visualization of the respective prediction map in QGIS.
}
\examples{


## -- create qml-file based on 'classaccur'-object:

# 1) get optimal classification scheme:

hsm<- HSMclass(refdata.gr, predictions.gr, nclasses = 6,
              iterations = 1000, coolfactor=0.99, InitTemp = 80,
              weight.norefs = 2, weight.classwidth = 2)

# 2) evaluate classification scheme:
acc.opti<- classAccuracy(refdata.gr, predictions.gr, def.int = hsm$best.classbreaks)

# 3) write qml-file for classification scheme:
\dontrun{
create_qml(acc.opti, out.file = "opti_class.qml", pallete.name = "YlOrRd")
}

}
