% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optcrit.R
\name{optcrit}
\alias{optcrit}
\alias{optcrit.summary.classaccur}
\title{optcrit}
\usage{
optcrit(object, ...)

\method{optcrit}{summary.classaccur}(object, ...)
}
\arguments{
\item{object}{object of class \code{'summary.classaccur'} containing an evaluated classification scheme.}

\item{...}{additional arguments, so far ignored.}
}
\description{
Displays the values of the 3 optimization criteria of \code{\link{HSMclass}}
}
