% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hsm.R
\name{plot.hsmclass}
\alias{plot.hsmclass}
\title{Plotting objects of class \code{'hsmclass'}}
\usage{
\method{plot}{hsmclass}(x, ...)
}
\arguments{
\item{x}{object of class \code{'hsmclass'} containing results of a
optimized classification scheme}

\item{...}{additional arguments, so far ignored.}
}
\description{
Function \code{plot.hsmclass} provides four plots that describe the
behaviour of the Simulated Annealing Heuristic applied by \code{\link{HSMclass}}.
The plots can be used to optimize the heuristics' behaviour by altering the optimization
parameters \code{coolfactor} and \code{InitTemp} in \code{\link{HSMclass}}.
}
\examples{

hsm<- HSMclass(refdata.gr, predictions.gr, nclasses = 6,
                 iterations = 1000, coolfactor=0.99, InitTemp = 80,
                 weight.norefs = 2, weight.classwidth = 2)
plot(hsm)

}
