% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_documentation.R
\docType{data}
\name{predictions.gr}
\alias{predictions.gr}
\title{A named vector containing timber volume predictions based on a multiple linear regression model
for each terrestrial observed timber volume stored in \code{\link{refdata.gr}}.}
\format{An object of class \code{numeric} of length 67.}
\usage{
predictions.gr
}
\description{
A named vector containing timber volume predictions based on a multiple linear regression model
for each terrestrial observed timber volume stored in \code{\link{refdata.gr}}.
}
\references{
Hill, A., Breschan, J., & Mandallaz, D. (2014). Accuracy assessment of timber
          volume maps using forest inventory data and LiDAR canopy height models.
          \emph{Forests}, 5(9), 2253-2275.
}
\keyword{datasets}
