% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata_documentation.R
\docType{data}
\name{refdata.gr}
\alias{refdata.gr}
\title{A named vector containing observations of the terrestrial timber volume at 67 systematically arranged sample plots.
Data is provided from a terrestrial forest inventory in the canton of Grisons, Switzerland in the year 2007}
\format{An object of class \code{numeric} of length 67.}
\source{
The terrestrial data are kindly provided by the forest service of the canton grisons.
}
\usage{
refdata.gr
}
\description{
A named vector containing observations of the terrestrial timber volume at 67 systematically arranged sample plots.
Data is provided from a terrestrial forest inventory in the canton of Grisons, Switzerland in the year 2007
}
\keyword{datasets}
