% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_classaccur.R
\name{summary.classaccur}
\alias{summary.classaccur}
\title{Summarizing  Evaluation of Classification Accuracy}
\usage{
\method{summary}{classaccur}(object, ...)
}
\arguments{
\item{object}{object of class \code{'classaccur'}}

\item{...}{additional arguments, so far ignored.}
}
\value{
\code{summary.classaccur} returns an object of class \code{"summary.classaccur"}.

An object of class \code{"summary.classaccur"} returns a \code{list} of the following components:

\item{accmat}{a \code{\link[base]{data.frame}} summarizing the user's- and producer's accuracy}
\item{rsquared_val}{the coefficient of determination of the prediction model, calulated based on
                     \code{refdata} and \code{predictions}}
\item{predictions}{ \code{vector} containing the predictions given to \code{classAccuracy} }
\item{overall.accuracy}{the overall accuracy of the classification scheme}
\item{conf.oaa}{ \code{vector} containing \code{[1]} the lower confidence limit of the OAA,
                                           \code{[2]} the upper confidence limit of the OAA,
                                       and \code{[3]} the confidence level}
 \item{cohenskappa}{Cohen's Kappa Coefficient}
 \item{Quantity.Disagreement}{ the quantity disagreement}
 \item{Allocation.Disagreement}{the allocation disagreement}
}
\description{
\code{summary} methods for class \code{'classaccur'}
}
\examples{

## -- Summarize classification accuracy:

acc.equal<- classAccuracy(refdata.gr, predictions.gr, equal.int = 100)
summ.<- summary(acc.equal)

# print summary-object:
summ.

# extract accuracy-data.frame:
summ.$accmat

}
