% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hsmclass.R
\name{summary.hsmclass}
\alias{summary.hsmclass}
\title{Summarizing Optimization of Classification Scheme}
\usage{
\method{summary}{hsmclass}(object, ...)
}
\arguments{
\item{object}{object of class \code{'hsmclass'}}

\item{...}{additional arguments, so far ignored.}
}
\value{
\code{summary.hsmclass} returns an object of class \code{"summary.hsmclass"}.

An object of class \code{"summary.hsmclass"} returns a \code{list} of the following components:

\item{call}{ the function call passed to function \code{HSMclass}}
\item{iterations}{number of iterations used in heuristic}
\item{bestever.iterationmode}{number of times the heuristic should be repeated.
                              \code{NA} indicates that option was not used.}
\item{classmat}{a \code{\link[base]{data.frame}} summarizing the identified optimal classification scheme}
}
\description{
\code{summary} methods for class \code{'hsmclass'}
}
\seealso{
\code{\link{HSMclass}}
}
